\name{PharmPow-package}
\alias{PharmPow-package}
\alias{PharmPow}
\docType{package}
\title{
Pharmacometric Power calculations for mixed study designs
}
\description{
This package contains functions performing power calculation for mixed (sparse/dense sampled) pharmacokinetic study designs. The input data for these functions is taylored for NONMEM .phi files.
}
\details{
\tabular{ll}{
Package: \tab PharmPow\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2013-09-16\cr
License: \tab MIT + file LICENSE\cr
}
PharmPow_parallel
PharmPow_crossover
fig_PharmPow_parallel
fig_PharmPow_crossover
}
\author{
Frank Kloprogge & Joel Tarning

Maintainer: Frank Kloprogge <frank@tropmedres.ac>
}
\keyword{ package }
